+++ unit/util/simplify_expr.cpp
@@ -16,7 +16,7 @@ Author: Michael Tautschnig
 #include <util/namespace.h>
 #include <util/pointer_predicates.h>
 #include <util/simplify_expr.h>
-#include <util/simplify_expr_class.h>
+#include <util/simplify_utils.h>
 #include <util/std_expr.h>
 #include <util/symbol_table.h>
 
@@ -88,24 +88,23 @@ TEST_CASE("expr2bits and bits2expr respect bit order", "[core][util]")
 {
   symbol_tablet symbol_table;
   namespacet ns(symbol_table);
-  simplify_exprt simp(ns);
 
   exprt deadbeef = from_integer(0xdeadbeef, unsignedbv_typet(32));
 
-  const auto le = simp.expr2bits(deadbeef, true);
+  const auto le = expr2bits(deadbeef, true, ns);
   REQUIRE(le.has_value());
   REQUIRE(le->size() == 32);
 
   const auto should_be_deadbeef1 =
-    simp.bits2expr(*le, unsignedbv_typet(32), true);
+    bits2expr(*le, unsignedbv_typet(32), true, ns);
   REQUIRE(deadbeef == *should_be_deadbeef1);
 
-  const auto be = simp.expr2bits(deadbeef, false);
+  const auto be = expr2bits(deadbeef, false, ns);
   REQUIRE(be.has_value());
   REQUIRE(be->size() == 32);
 
   const auto should_be_deadbeef2 =
-    simp.bits2expr(*be, unsignedbv_typet(32), false);
+    bits2expr(*be, unsignedbv_typet(32), false, ns);
   REQUIRE(deadbeef == *should_be_deadbeef2);
 }
